Move core networking routines into a utils module
